requirements.txt
# generated using pipreqs 
keras==3.6.0
matplotlib==3.8.1
mediapipe==0.10.14
numpy==2.1.2
opencv_contrib_python==4.10.0.84
opencv_python==4.8.1.78
scikit_learn==1.5.2
tensorflow_intel==2.17.0
